+++ calculation/output/DataFileWriter.h
@@ -26,7 +26,7 @@ class RecordBuffer {
       FieldValue                      & GetFieldValue(const std::string& sFieldName);
       FieldValue                      & GetFieldValue(unsigned int iFieldIndex);
       const FieldValue                & GetFieldValue(unsigned int iFieldIndex) const;
-      unsigned int                      GetNumFields() const { return gvFieldValues.size();}
+      size_t                            GetNumFields() const { return gvFieldValues.size();}
       void                              SetAllFieldsBlank(bool bBlank);
       void                              SetFieldIsBlank(const std::string& sFieldName, bool bBlank);
       void                              SetFieldIsBlank(unsigned int iFieldNumber, bool bBlank);
